Set default namespace for cljformat
When running a _tool_ via `:aliases`, you have to specify the namespace
of the function you're calling.

`clj -T:cljfmt cljfmt.tool/check`

Now, we have specified this default namespace in our `deps.edn` and we
can remove when invoking the command at the CLI or in our `Justfile`.
The default namespace was found by looking at the `:ns-default` keyword
in `deps.edn` in `cljfmt`'s GitHub repo (and it was also seen in some
examples, but not relevant to running via `:aliases`)

P.S. I would still like to know why this `:ns-default` (default
namespace) is not used if it's present in the tools own `deps.edn` and
is a documented standard.