18 Group route handlers by page
Refactors the route handlers by page/feature. The routes are moved into
namespaces based on the page they will serve.

We want as few routes as a possible to be used for serving multiple
pages. Ideally no route will serve multipe pages. If this step isn't
taken early, it can become a hairball.

The new route namespaces expose function that returns a collection of
routes (with handlers), rather than needing to build up a tree of routes
from accross the different namespaces.

So each page (namespace? feature?) and it's routes are all bundled
together neatly.